% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{add_mutations}
\alias{add_mutations}
\alias{add_substitution}
\alias{add_insertion}
\alias{add_deletion}
\title{Add mutations manually from R.}
\usage{
add_substitution(var_set_, var_ind, seq_ind, nucleo_, new_pos_)

add_insertion(var_set_, var_ind, seq_ind, nucleos_, new_pos_)

add_deletion(var_set_, var_ind, seq_ind, size_, new_pos_)
}
\arguments{
\item{var_set_}{External pointer to a C++ \code{VarSet} object}

\item{var_ind}{Integer index to the desired variant. Uses 0-based indexing!}

\item{seq_ind}{Integer index to the desired sequence. Uses 0-based indexing!}

\item{nucleo_}{Character to substitute for existing one.}

\item{new_pos_}{Integer index to the desired subsitution location.
Uses 0-based indexing!}

\item{nucleos_}{Nucleotides to insert at the desired location.}

\item{size_}{Size of deletion.}
}
\description{
Note that all indices are in 0-based C++ indexing. This means that the first
item is indexed by \code{0}, and so forth.
}
\section{Functions}{
\itemize{
\item \code{add_substitution}: Add a substitution.

\item \code{add_insertion}: Add an insertion.

\item \code{add_deletion}: Add a deletion.
}}

