% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{binding_sites}
\alias{binding_sites}
\title{Binding sites for selected restriction enzymes.}
\format{A list of length 11. For each element in the list...
\describe{
\item{name}{The element's name is the restriction enzyme's name.
The enzymes present in this list are
\emph{AclI}, \emph{ApeKI}, \emph{AscI}, \emph{BspEI}, \emph{BstBI},
\emph{EcoT22I}, \emph{FspI}, \emph{MluI-HF}, \emph{NruI-HF}, \emph{PstI},
and \emph{SbfI}.}
\item{sites}{Enzyme binding site sequences. This vector is of the binding sites,
5' then 3', for each unique site that the enzyme can bind to.}
}}
\source{
\url{http://www.neb.com}
}
\usage{
binding_sites
}
\description{
Binding sites for selected restriction enzymes.
}
\keyword{datasets}
