% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_genome}
\alias{create_genome}
\title{Create \code{RefGenome} pointer based on nucleotide equilibrium frequencies.}
\usage{
create_genome(n_seqs, len_mean, len_sd = 0, pi_tcag = numeric(0),
  n_cores = 1L)
}
\arguments{
\item{n_seqs}{Number of sequences.}

\item{len_mean}{Mean for the gamma distribution for sequence sizes.}

\item{len_sd}{Standard deviation for the gamma distribution for sequence sizes.
If set to \code{<= 0}, all sequences will be the same length. Defaults to \code{0}.}

\item{pi_tcag}{Vector of nucleotide equilibrium frequencies for
"T", "C", "A", and "G", respectively. Defaults to \code{rep(0.25, 4)}.}

\item{n_cores}{Number of cores to use via OpenMP.}
}
\value{
External pointer to a \code{RefGenome} C++ object.
}
\description{
Function to create random sequences for a new reference genome object.
}
\details{
Note that this function will never return empty sequences.
}
\examples{

genome <- create_genome(10, 100e3, 100, pi_tcag = c(0.1, 0.2, 0.3, 0.4))

}
