% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digest.R
\name{digest}
\alias{digest}
\title{Digest reference genome or variants from a genome.}
\usage{
digest(object, enzyme_names, custom_enzymes, chunk_size = 0, n_cores = 1)
}
\arguments{
\item{object}{Either a \code{ref_genome} or \code{variants} object.}

\item{enzyme_names}{Name of enzymes that should be present inside the package data
object \code{binding_sites} (in the \code{enzyme} column).}

\item{custom_enzymes}{Custom enzymes for those not present in internal data.}

\item{chunk_size}{The size of chunks to break up sqeuences into when digesting
a \code{ref_genome} or \code{variants} object.
Changing this might affect performance, for better or worse.
The default worked best on my computer. Defaults to \code{1000}.}

\item{n_cores}{Number of cores to use for parallel processing. This argument is
ignored if OpenMP is not enabled. Defaults to \code{1}.}
}
\value{
All changes occur in place, but the input object is return invisibly
so that piping works.
}
\description{
\emph{Note:} This will override any digestions currently in place in the
object. If you want to add a new digestion, re-run this function with the names
of all enzymes you're interested in included in the \code{enzyme_names} argument.
}
