% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haps_functions.R
\name{haps_theta}
\alias{haps_theta}
\title{Organize information to create haplotypes using theta parameter}
\usage{
haps_theta(theta, n_haps)
}
\arguments{
\item{theta}{Population-scaled mutation rate.}

\item{n_haps}{Number of desired haplotypes.}
}
\value{
A \code{haps_theta_info} object containing information used in \code{create_haplotypes}
to create variant haplotypes.
This class is just a wrapper around a list containing the phylogenetic tree
and \code{theta} parameter, which you can view (but not change) using the object's
\code{phylo()} and \code{theta()} methods, respectively.
}
\description{
This function organizes higher-level information for creating haplotypes from
the population-scaled mutation rate and a desired number of haplotypes.
}
