% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illumina.R
\name{illumina}
\alias{illumina}
\title{Create and write Illumina reads to FASTQ file(s).}
\source{
Huang, W., L. Li, J. R. Myers, and G. T. Marth. 2012. ART: a next-generation
sequencing read simulator. \emph{Bioinformatics} \strong{28}:593–594.
}
\usage{
illumina(seq_object, out_prefix, n_reads, read_length, paired, frag_mean,
  frag_sd, matepair = FALSE, seq_sys = NULL, profile1 = NULL,
  profile2 = NULL, ins_prob1 = 9e-05, del_prob1 = 0.00011,
  ins_prob2 = 0.00015, del_prob2 = 0.00023, frag_len_min = NULL,
  frag_len_max = NULL, variant_probs = NULL, barcodes = NULL,
  prob_dup = 0.02, compress = FALSE, n_cores = 1L,
  read_chunk_size = 100L, show_progress = FALSE)
}
\arguments{
\item{seq_object}{Sequencing object of class \code{ref_genome} or \code{variants}.}

\item{out_prefix}{Prefix for the output file(s), including entire path except
for the file extension.}

\item{n_reads}{Number of reads you want to create.}

\item{read_length}{Length of reads.}

\item{paired}{Logical for whether to use paired-end reads.}

\item{frag_mean}{Mean of the Gamma distribution that generates fragment sizes.}

\item{frag_sd}{Standard deviation of the Gamma distribution that generates
fragment sizes.}

\item{matepair}{Logical for whether to simulate mate-pair reads.
This argument is ignored if \code{paired} is \code{FALSE}.
Defaults to \code{FALSE}.}

\item{seq_sys}{Full or abbreviated name of sequencing system to use.
See "Sequencing systems" section for options.
See "Sequencing profiles" section for more information on how this argument,
\code{profile1}, and \code{profile2} are used to specify profiles.
Defaults to \code{NULL}.}

\item{profile1}{Custom profile file for read 1.
See "Sequencing profiles" section for more information on how this argument,
\code{profile2}, and \code{seq_sys} are used to specify profiles.
Defaults to \code{NULL}.}

\item{profile2}{Custom profile file for read 2.
See "Sequencing profiles" section for more information on how this argument,
\code{profile1}, and \code{seq_sys} are used to specify profiles.
Defaults to \code{NULL}.}

\item{ins_prob1}{Insertion probability for read 1. Defaults to \code{0.00009}.}

\item{del_prob1}{Deletion probability for read 1. Defaults to \code{0.00011}.}

\item{ins_prob2}{Insertion probability for read 2. Defaults to \code{0.00015}.}

\item{del_prob2}{Deletion probability for read 2. Defaults to \code{0.00023}.}

\item{frag_len_min}{Minimum fragment size. A \code{NULL} value results in the read length.
Defaults to \code{NULL}.}

\item{frag_len_max}{Maximum fragment size.
A \code{NULL} value results in \code{2^32-1}, the maximum allowed value.
Defaults to \code{NULL}}

\item{variant_probs}{Relative probability of sampling each variant.
This is ignored if sequencing a reference genome.
\code{NULL} results in all having the same probability.
Defaults to \code{NULL}.}

\item{barcodes}{Character vector of barcodes for each variant, or a single barcode
if sequencing a reference genome. \code{NULL} results in no barcodes.
Defaults to \code{NULL}.}

\item{prob_dup}{A single number indicating the probability of duplicates.
Defaults to \code{0.02}.}

\item{compress}{A logical for whether to compress output FASTQ files using gzip.
Defaults to \code{FALSE}.}

\item{n_cores}{The number of cores to use in processing.
This argument is ignored if the package was not compiled with OpenMP.
Defaults to \code{1}.}

\item{read_chunk_size}{The number of reads to store before writing to disk.
Increasing this number should improve speed but take up more memory.
Defaults to \code{100}.}

\item{show_progress}{Logical for whether to show a progress bar.
Defaults to \code{FALSE}.}
}
\value{
Nothing is returned.
}
\description{
From either a reference genome or set of haploid variants, create Illumina reads
from error profiles and write them to FASTQ output file(s).
}
\section{Sequencing profiles}{

This section outlines how to use the \code{seq_sys}, \code{profile1},
and \code{profile2} arguments.
If all arguments are \code{NULL} (their defaults), a sequencing system is chosen
based on the read length.
If, however, one or more arguments has been provided, then how they're provided
should depend on whether you want single- or paired-end reads.

\strong{For single-end reads}
\itemize{
\item \code{profile2} should be \code{NULL}.
\item Only \code{seq_sys} or \code{profile1} should be provided, not both.
}

\strong{For paired-end reads}
\itemize{
\item If providing \code{seq_sys}, don't provide either \code{profile1} or \code{profile2}.
\item If providing \code{profile1}, you must also provide \code{profile2} (they can be the
same if you want) and you cannot provide \code{seq_sys}.
}
}

\section{Sequencing systems}{

Sequencing system options are the following, where, for each system,
"name" is the full name, "abbrev" is the abbreviated name,
"max_len" indicates the maximum allowed read length,
and
"paired" indicates whether paired-end sequencing is allowed.

| name                 | abbrev  | max_len | paired |
|:---------------------|:--------|:--------|:-------|
| Genome Analyzer I    | GA1     | 44      | Yes    |
| Genome Analyzer II   | GA2     | 75      | Yes    |
| HiSeq 1000           | HS10    | 100     | Yes    |
| HiSeq 2000           | HS20    | 100     | Yes    |
| HiSeq 2500           | HS25    | 150     | Yes    |
| HiSeqX v2.5 PCR free | HSXn    | 150     | Yes    |
| HiSeqX v2.5 TruSeq   | HSXt    | 150     | Yes    |
| MiniSeq TruSeq       | MinS    | 50      | No     |
| MiSeq v1             | MSv1    | 250     | Yes    |
| MiSeq v3             | MSv3    | 250     | Yes    |
| NextSeq 500 v2       | NS50    | 75      | Yes    |
}

\section{ID lines}{

The ID lines for FASTQ files are formatted as such:

\code{@<genome name>-<sequence name>-<starting position>[/<read#>]}

where the part in \code{[]} is only for paired-end Illumina reads, and where \code{genome name}
is always \code{REF} for reference genomes (as opposed to variants).
}

\examples{
\dontrun{
rg <- create_genome(10, 100e3, 100)
illumina(rg, "illumina_reads", n_reads = 100,
         read_length = 100, paired = FALSE,
         frag_mean = 400, frag_sd = 100)
}

}
