% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecular_evolution.R
\name{make_mevo}
\alias{make_mevo}
\title{Make a \code{mevo} object to store information needed for molecular evolution simulation.}
\usage{
make_mevo(reference, sub, ins = NULL, del = NULL, site_var = NULL,
  chunk_size = 100)
}
\arguments{
\item{sub}{A list containing the parameters for substitutions.
The \code{model} field is always required within this list, and this specifies
the model used for substitutions.
It takes one of the following options:
\code{"JC69"}, \code{"K80"}, \code{"F81"}, \code{"HKY85"}, \code{"TN93"}, \code{"F84"}, \code{"GTR"}, or \code{"UNREST"}.
See \code{vignette("sub-models")} for more information on these models and
their required parameters.}

\item{ins}{A list containing the parameters for insertions.
Passing \code{NULL} to this argument results in no insertions.
See "Indels" section for more information.
Defaults to \code{NULL}.}

\item{del}{A list containing the parameters for deletions.
Passing \code{NULL} to this argument results in no deletions.
See "Indels" section for more information.
Defaults to \code{NULL}.}

\item{site_var}{List of parameters for generating variability in mutation
rates among sites (for both substitutions and indels).
Passing \code{NULL} to this argument results in no variability among sites.
See "Site variation" section for more information.
Defaults to \code{NULL}.}

\item{chunk_size}{The size of "chunks" of sequences to first sample uniformly
before doing weighted sampling by rates for each sequence location.
Uniformly sampling before doing weighted sampling dramatically speeds up
the mutation process (especially for very long sequences) and has little
effect on the sampling probabilities.
Higher values will more closely resemble sampling without the uniform-sampling
step, but will be slower.
Set this to \code{0} to not uniformly sample first.
From testing on a chromosome of length \code{1e6}, a \code{chunk_size} value of \code{100}
offers a ~10x speed increase and doesn't differ significantly from sampling
without the uniform-sampling step.
Defaults to \code{100}.}
}
\value{
An object of class \code{\link{mevo}}.
}
\description{
Make a \code{mevo} object to store information needed for molecular evolution simulation.
}
\section{Indels}{

Both insertions and deletions require the \code{rate} parameter, which specifies
the overall insertion/deletion rate among all lengths.
The \code{rate} parameter is ultimately combined with a vector of relative rates among
the different lengths of insertions/deletions from 1 to the maximum
possible length.
There are three different ways to specify/generate relative-rate values.
\enumerate{
\item Assume that rates are proportional to \code{exp(-L)} for indel length
\code{L} from 1 to the maximum length (Albers et al. 2011).
This method will be used if the following arguments are provided:
\itemize{
\item \code{rate}
\item \code{max_length}
}
\item Generate relative rates from a Lavalette distribution
(Fletcher and Yang 2009), where the rate for length \code{L} is proportional to
\code{{L * max_size / (max_size - L + 1)}^(-a)}.
This method will be used if the following arguments are provided:
\itemize{
\item \code{rate}
\item \code{max_length}
\item \code{a}
}
\item Directly specify values by providing a numeric vector of relative
rates for each insertion/deletion length from 1 to the maximum length.
This method will be used if the following arguments are provided:
\itemize{
\item \code{rate}
\item \code{rel_rates}
}
}
Note that providing extra parameter(s) will return an error.
}

\section{Site variation}{

A site's deviance from the average mutation rate is determined by its
"gamma distance".
A site's overall mutation rate is the mutation rate for that nucleotide
(substitution + indel) multiplied by the site's gamma distance.
There are two options for specifying gamma distances:
\enumerate{
\item Generate gamma distances from a Gamma distribution.
This option requires the following arguments:
\describe{
\item{\code{shape}}{Shape parameter for the Gamma distribution,
where the variance of the distribution is \code{1 / shape}.
The mean is fixed to 1.}
\item{\code{region_size}}{Size of regions to break the genome into,
where all sites within a region have the same gamma distance.}
}
\item Manually input matrices that specify the gamma distance and end points
for regions each gamma distances refers to.
This option requires the following argument:
\describe{
\item{\code{mats}}{List of matrices, one for each sequence in the genome.
Each matrix should have two columns.
The first should contain the end points for each region.
The second should contain the gamma distances for each region.
Note that if gamma distances don't have a mean (weighted by
sequence length for each gamma-distance value) equal to 1,
you're essentially changing the overall mutation rate.}
}
}
}

\examples{
ref <- create_genome(10, 1000)
mevo <- make_mevo(ref, list(model = "JC69", lambda = 1))

}
\references{
Albers, C. A., G. Lunter, D. G. MacArthur, G. McVean, W. H. Ouwehand, and R. Durbin.
2011. Dindel: accurate indel calls from short-read data. Genome Research 21:961–973.

Fletcher, W., and Z. Yang. 2009. INDELible: a flexible simulator of
biological sequence evolution. Molecular Biology and Evolution 26:1879–1888.
}
