% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\docType{class}
\name{mevo}
\alias{mevo}
\title{An R6 class containing information needed for molecular evolution.}
\format{An \code{\link[R6]{R6Class}} generator object}
\usage{
mevo
}
\value{
An object of class \code{mevo}.
}
\description{
You shouldn't need to interact with this class much, if at all.
This class is designed to be made in \code{create_mevo}, then to store information for use
in \code{create_variants}.
}
\section{Fields}{

\describe{
\item{\code{Q}}{A matrix of substitution rates for each nucleotide.}

\item{\code{pi_tcag}}{Vector of nucleotide equilibrium frequencies for "T", "C", "A", and
"G", respectively.}

\item{\code{insertion_rates}}{Vector of insertion rates by length.}

\item{\code{deletion_rates}}{Vector of deletion rates by length.}

\item{\code{gamma_mats}}{List of matrices specifying "gamma distances" (see definition in
\code{?create_mevo}) for each sequence.}

\item{\code{chunk_size}}{The size of "chunks" of sequences to first sample uniformly
before doing weighted sampling by rates for each sequence location.
See \code{?create_mevo} for more information.}
}}

\section{Methods}{

\describe{
\item{\code{mu()}}{Calculates the average overall mutation rate at equilibrium.}
\item{\code{q()}}{Calculates the mutation rate for each nucleotide.}
}
}

\seealso{
\code{\link{create_mevo}}
}
\keyword{datasets}
