% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nucleobase_legend}
\alias{nucleobase_legend}
\title{Legend for the single-letter code of nucleobases indicating restriction sequences.}
\format{A data frame of 28 rows and two columns:
\describe{
\item{code}{The letter indicating more than one possible nucleotides.}
\item{nucleotides}{One of the multiple nucleotides the code can refer to.}
}}
\source{
\url{https://en.wikipedia.org/wiki/List_of_restriction_enzyme_cutting_sites:_A#Whole_list_navigation}
}
\usage{
nucleobase_legend
}
\description{
Legend for the single-letter code of nucleobases indicating restriction sequences.
}
\keyword{datasets}
