% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_variants.R
\name{random_variants}
\alias{random_variants}
\title{Construct a \code{VarSet} object from a reference genome and summary statistics.}
\usage{
random_variants(dna_set_in, n_vars, theta_w, theta_pi,
  indel_probs = exp(-1:-10), snp_probs = rep(0.25, 4),
  snp_proportion = NULL, n_cores = 1, n2N = 50, alpha = 0.8)
}
\arguments{
\item{dna_set_in}{A \code{dna_set} object of sequences representing the reference
genome.}

\item{n_vars}{The number of variants to create.}

\item{theta_w}{Watterson's estimator for the focal population.}

\item{theta_pi}{Average nucleotide diversity for the focal population.}

\item{indel_probs}{Relative probabilities of indel types and sizes.
If insertions and deletions have the same probabilities, this is simply
a numeric vector where the value in location \code{i} indicates the relative
probability of an indel of size \code{i}.
If insertions and deletions do not have the same probabilities, then
this argument should be a list where the \code{insertion} and \code{deletion} fields are
numeric vectors specifying relative probabilities for insertions and deletions,
respectively.
Note that if specifying a list, the proportion of insertions to deletions will
be \code{sum(indel_probs$insertions) / sum(indel_probs$deletions)}.
Defaults to \code{exp(-1:-10)}.}

\item{snp_probs}{Relative probabilities of substitution types:
"A", "C", "G", and "T" respectively. Defaults to \code{rep(0.25, 4)}.}

\item{snp_proportion}{The proportion of mutations (not sites) that are SNPs.
Defaults to the proportion calculated using the average ratio of indels
to substitutions in eukaryotes from Sung et al. (2016).}

\item{n_cores}{Number of cores to use. Defaults to 1.}

\item{n2N}{A numeric threshold placed on the algorithm used to find new locations.
This is not recommended to be changed. Defaults to 50.}

\item{alpha}{A numeric threshold placed on the algorithm used to find new locations.
This is not recommended to be changed. Defaults to 0.8.}
}
\value{
A \code{VarSet} object.
}
\description{
This function creates a \code{VarSet} object, which is designed to be a
low-memory way to store variants.
(The variants class also prevents what might otherwise be an annoyingly long list
from ever printing in the console.)
}
\examples{
\dontrun{
n_vars <- 10
dna_set_in <- dna_set$new(rando_seqs(100, 100))
set.seed(1)
varseq_out <- random_variants(dna_set_in, n_vars,
                              theta_w = 0.0045, theta_pi = 0.005)
}

}
\references{
Sung, W., M. S. Ackerman, M. M. Dillon, T. G. Platt, C. Fuqua, V. S. Cooper, and M.
Lynch. 2016.
Evolution of the insertion-deletion mutation rate across the tree of life.
\emph{G3: Genes|Genomes|Genetics} \strong{6}:2583-2591.
}
