% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{read_fasta}
\alias{read_fasta}
\title{Read a fasta file.}
\usage{
read_fasta(fasta_file, fai_file = NULL, cut_names = TRUE,
  rm_soft_mask = TRUE)
}
\arguments{
\item{fasta_file}{File name of the fasta file.}

\item{fai_file}{File name of the fasta index file.
Providing this argument speeds up the reading process significantly.
Defaults to \code{NULL}, which indicates the fasta file is not indexed.}

\item{cut_names}{Boolean for whether to cut sequence names at the first space.
This argument is ignored if \code{fai_file} is not \code{NULL}.
Defaults to \code{TRUE}.}

\item{rm_soft_mask}{Boolean for whether to remove soft-masking by making
sequences all uppercase. Defaults to \code{TRUE}.}
}
\value{
A \code{\link{ref_genome}} object.
}
\description{
Accepts uncompressed and gzipped fasta files.
}
