% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\docType{class}
\name{ref_genome}
\alias{ref_genome}
\title{An R6 class representing a reference genome.}
\format{An \code{\link[R6]{R6Class}} generator object}
\usage{
ref_genome
}
\value{
An object of class \code{ref_genome}.
}
\description{
\emph{Note:} Do NOT change fields in this class directly.
It will cause your R session to do bad things.
(Ever seen the bomb popup on RStudio? Manually mess with these fields and you
surely will.)
}
\section{Fields}{

\describe{
\item{\code{genome}}{An \code{externalptr} to a C++ object storing the sequences
representing the genome.}

\item{\code{digests}}{An \code{externalptr} to a C++ object storing the digestion of the
genome, if a digestion has been carried out. It's \code{NULL} otherwise.}
}}

\seealso{
\code{\link{read_fasta}} \code{\link{create_genome}}
}
\keyword{datasets}
