% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{table_gammas}
\alias{table_gammas}
\title{Faster version of table function to count the number of mutations in Gamma regions.}
\usage{
table_gammas(gamma_ends, positions)
}
\arguments{
\item{gamma_ends}{Vector of endpoints for gamma regions}

\item{positions}{Vector of positions that you want to bin into gamma regions.}
}
\description{
Faster version of table function to count the number of mutations in Gamma regions.
}
