% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{write_vcf}
\alias{write_vcf}
\title{Write variant info from a \code{variants} object to a VCF file.}
\usage{
write_vcf(vars, out_prefix, compress = FALSE, sample_matrix = NULL)
}
\arguments{
\item{vars}{A \code{variants} object.}

\item{out_prefix}{Prefix for the output file VCF.}

\item{compress}{Boolean for whether to compress using \code{"gzip"}.
Defaults to \code{FALSE}.}

\item{sample_matrix}{Matrix to specify how haploid variants are grouped into samples
if samples are not haploid. There should be one row for each sample, and
each row should contain indices or names for the variants present in that sample.
Indices/names for variants can be repeated across and within rows.
The number of columns indicates the ploidy level: 2 columns for diploid,
3 for triploid, 4 for tetraploid, and so on;
there is no limit to the ploidy level.
If this argument is \code{NULL}, it's assumed that each variant is its own
separate sample.
Defaults to \code{NULL}.}
}
\value{
\code{NULL}
}
\description{
Write variant info from a \code{variants} object to a VCF file.
}
