% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{binding_sites}
\alias{binding_sites}
\title{Binding sites for selected restriction enzymes.}
\format{A 695 x 2 data frame with the following columns:
\describe{
\item{enzyme}{Restriction enzyme name.}
\item{sites}{Enzyme binding site sequences.
See \code{\link{nucleobase_legend}} for what bases other than \code{T}, \code{C}, \code{A},
and \code{G} mean.
Each \code{/} indicates a cleavage point.
According to NEB (see link below under Source):
"Numbers in parentheses indicate point of cleaveage for non-palindromic
enzymes."
These types of enzymes are not implemented.}
}}
\source{
\url{https://www.neb.com/tools-and-resources/selection-charts/alphabetized-list-of-recognition-specificities}
}
\usage{
binding_sites
}
\description{
Binding sites for selected restriction enzymes.
}
\keyword{datasets}
