% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_variants.R
\name{create_variants}
\alias{create_variants}
\title{Create variants from a reference genome.}
\usage{
create_variants(reference, method, method_info, mevo_obj, n_cores = 1,
  show_progress = FALSE)
}
\arguments{
\item{reference}{A \code{ref_genome} object from which to generate variants.
This argument is required.}

\item{method}{Method to use for generating variants.
Options are as follows:
\describe{
\item{\code{"phylo"}}{a single phylogenetic tree from a \code{phylo} object.}
\item{\code{"coal_obj"}}{a coalescent-simulator object from the \code{scrm} or \code{coala}
packages.}
\item{\code{"ms_file"}}{a file containing output from a coalescent simulator in the
format of the \code{ms} program.}
\item{\code{"newick"}}{a NEWICK file containing a phylogenetic tree.}
\item{\code{"theta"}}{an estimate for theta, the population-scaled mutation rate.}
\item{\code{"vcf"}}{a variant call format (VCF) file that directly specifies
variants. This method does not work if the \code{vcfR} package isn't installed.
\strong{\emph{NOTE:}} If this method is chosen, all arguments other than
\code{reference}, \code{method}, and \code{method_info} are ignored.}
}}

\item{method_info}{Object containing information used for the given method.
See "Method arguments" section for which arguments are used for each method.}

\item{mevo_obj}{A \code{mevo} object that stores molecular-evolution information.
See \code{\link{make_mevo}} for more information.}

\item{n_cores}{Number of cores to use for parallel processing.
This argument is ignored if OpenMP is not enabled.
Cores are spread across sequences, so it
doesn't make sense to supply more cores than sequences in the reference genome.
Defaults to \code{1}.}

\item{show_progress}{Boolean for whether to show a progress bar during processing.
Defaults to \code{FALSE}.}
}
\description{
Create variants from a reference genome.
}
\section{Method arguments}{

Below, I describe what the \code{method_info} should look like for each possible method.
\describe{
\item{\code{method = "phylo"}}{A single \code{\link[ape]{phylo}} object.}
\item{\code{method = "coal_obj"}}{A single \code{list} with a \code{trees} field inside.
For what this \code{trees} field should look like, see output from the
\code{scrm} or \code{coala} package. (These packages are not required to be installed
when installing \code{gemino}.)}
\item{\code{method = "ms_file"}}{A single string specifying the name of the file
containing the \code{ms}-style coalescent output.}
\item{\code{method = "newick"}}{A single string specifying the name of the NEWICK file
containing the phylogeny.}
\item{\code{method = "theta"}}{A named vector or list containing the fields \code{theta}
and \code{n_vars}, specifying the theta parameter (population-scaled mutation rate)
and number of desired variants, respectively.}
\item{\code{method = "vcf"}}{A single string specifying the name of the VCF file.}
}
}

