% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_variants.R
\name{create_variants}
\alias{create_variants}
\title{Create variants from a reference genome.}
\usage{
create_variants(reference, n_variants, phylo_control, vcf_file, random_control,
  n_cores = 1)
}
\arguments{
\item{reference}{A \code{ref_genome} object from which to generate variants.}

\item{n_variants}{The number of haploid variants to create.}

\item{phylo_control}{A list containing parameters for phylogenetic evolution.}

\item{vcf_file}{The file name for a VCF file to read from.}

\item{random_control}{A list containing parameters for random mutation addition.}

\item{n_cores}{Number of cores to use for parallel processing. This argument is
ignored if OpenMP is not enabled. Defaults to \code{1}.}
}
\description{
Create variants from a reference genome.
}
