% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackal.R
\docType{package}
\name{jackal}
\alias{jackal}
\alias{jackal-package}
\title{jackal: An efficient, flexible molecular evolution and sequencing simulator.}
\description{
\code{jackal} efficiently (i) reads and simulates reference genomes;
(ii) generates variants using summary statistics, phylogenies, Variant
Call Format (VCF) files, and coalescent simulations—the latter of which can include
selection, recombination, and demographic fluctuations;
(iii) simulates sequencing error, mapping qualities, and optical/PCR duplicates; and
(iv) writes outputs to standard file formats.
\code{jackal} can simulate single, paired-end, or mate-pair Illumina reads, as well as
reads from Pacific BioSciences.
}
