% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\docType{class}
\name{variants}
\alias{variants}
\title{An R6 class representing haploid variants from a reference genome}
\format{An \code{\link[R6]{R6Class}} generator object}
\usage{
variants
}
\value{
An object of class \code{variants}.
}
\description{
Note: Do NOT change fields in this class directly. It will likely cause your
R session to do bad things.
}
\section{Fields}{

\describe{
\item{\code{genome}}{An \code{externalptr} to a C++ object storing the sequences
representing the genome.}

\item{\code{digests}}{An \code{externalptr} to a C++ object storing the digestion of the
genome, if a digestion has been carried out. It's \code{NULL} otherwise.}
}}

\keyword{datasets}
