% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars_functions.R
\name{vars_theta}
\alias{vars_theta}
\title{Create necessary information to create variants using theta parameter}
\usage{
vars_theta(theta, n_vars)
}
\arguments{
\item{theta}{Population-scaled mutation rate.}

\item{n_vars}{Mumber of desired variants.}
}
\value{
A \code{vars_theta_info} object containing information used in \code{create_variants}
to create haploid variants.
This class is just a wrapper around a list containing the phylogenetic tree
and \code{theta} parameter.
}
\description{
This function organizes higher-level information for creating variants from
the population-scaled mutation rate and a desired number of variants.
}
