% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{write_fasta}
\alias{write_fasta}
\title{Write a \code{ref_genome} object to a FASTA file.}
\usage{
write_fasta(seq_obj, out_prefix, compress = FALSE, text_width = 80)
}
\arguments{
\item{seq_obj}{A \code{ref_genome} or \code{variants} object.}

\item{out_prefix}{Prefix for the output file.}

\item{compress}{Boolean for whether to compress using \code{"gzip"}.
Defaults to \code{FALSE}.}

\item{text_width}{The number of characters per line in the output fasta file.
Defaults to \code{80}.}
}
\value{
\code{NULL}
}
\description{
Write a \code{ref_genome} object to a FASTA file.
}
