% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{write_fasta}
\alias{write_fasta}
\title{Write a \code{ref_genome} object to a FASTA file.}
\usage{
write_fasta(reference, file_name, text_width = 80, compress = FALSE)
}
\arguments{
\item{reference}{A \code{ref_genome} object.}

\item{file_name}{File name of the output fasta file.}

\item{text_width}{The number of characters per line in the output fasta file.
Defaults to \code{80}.}

\item{compress}{Boolean for whether to compress using \code{"gzip"}.
Defaults to \code{FALSE}.}
}
\value{
\code{NULL}
}
\description{
Write a \code{ref_genome} object to a FASTA file.
}
