% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{write_fasta}
\alias{write_fasta}
\title{Write to FASTA file.}
\usage{
write_fasta(file_name, ptr, text_width = 80, compression = "none")
}
\arguments{
\item{file_name}{File name of the output fasta file.}

\item{ptr}{External pointer to a \code{RefGenome} C++ object.}

\item{text_width}{The number of characters per line in the output fasta file.
Defaults to 80.}

\item{compression}{Type of compression. Takes either \code{"none"} or \code{"gzip"}.
Defaults to \code{"none"}.}
}
\value{
Nothing.
}
\description{
Write to FASTA file.
}
